+++ src/components/Crosshair_bak.vue
@@ -1,7 +1,7 @@
 <script>
 
-import Crosshair from './js/crosshair.js'
-import Utils from '../stuff/utils.js'
+import Crosshair from './js/crosshair3.js'
+// import Utils from '../stuff/utils.js'
 
 export default {
     name: 'Crosshair',