+++ src/components/overlays/Candles.vue
@@ -2,109 +2,87 @@
 // Renedrer for candlesticks + volume (optional)
 // It can be used as the main chart or an indicator
 
-import Overlay from '../../mixins/overlay.js'
+import Overlay from '../../mixins/overlay.js' // Where is it usage ?
+
 import { layout_cnv } from '../js/layout_cnv.js'
 import Candle from '../primitives/candle.js'
 import Volbar from '../primitives/volbar.js'
 import Price from '../primitives/price.js'
+import {computed, ref} from 'vue'
 
-export default {
-    name: 'Candles',
-    mixins: [Overlay],
-    methods: {
-        meta_info() {
-            return { author: 'C451', version: '1.2.1' }
-        },
-        init() {
-            this.price = new Price(this)
-        },
-        draw(ctx) {
+// props from parent
+const props = defineProps({
+  sub: Object,
+  data:Object,
+  layout:Object, // candle, volume
+  settings:Object,
+  colors:Object,
+})
 
-            // If data === main candlestick data
-            // render as main chart:
-            if (this.$props.sub === this.$props.data) {
-                var cnv = {
-                    candles: this.$props.layout.candles,
-                    volume: this.$props.layout.volume,
-                }
-            // Else, as offchart / onchart indicator:
-            } else {
-                cnv = layout_cnv(this)
-            }
+// internal reactive state var (deep)
+// Price object
+const priceObj = {price:{}};
+const priceState = ref(priceObj);
 
-            if (this.show_volume) {
-                var cv = cnv.volume
-                for (var i = 0, n = cv.length; i < n; i++) {
-                    new Volbar(this, ctx, cv[i])
-                }
-            }
-
-            var cc = cnv.candles
-            for (var i = 0, n = cc.length; i < n; i++) {
-                new Candle(this, ctx, cc[i])
-            }
-
-            if (this.price_line) this.price.draw(ctx)
-        },
-        use_for() { return ['Candles'] },
-
-        // In case it's added as offchart overlay
-        y_range() {
+// methods
+const meta_info = () => {return { author: 'C451', version: '1.2.1' }} // where ?
+const init = () => {priceState.value = new Price(priceObj)} // init value through method, can be shorten
+const use_for = () => { return ['Candles'] }
+const y_range = () =>  {
             var hi = -Infinity, lo = Infinity
-            for (var i = 0, n = this.sub.length; i < n; i++) {
-                let x = this.sub[i]
+            for (var i = 0, n = props.sub.length; i < n; i++) {
+                let x = props.sub[i]
                 if (x[2] > hi) hi = x[2]
                 if (x[3] < lo) lo = x[3]
             }
             return [hi, lo]
         }
-    },
 
-    // Define internal setting & constants here
-    computed: {
-        sett() {
-            return this.$props.settings
-        },
-        show_volume() {
-            return 'showVolume' in this.sett ?
-                this.sett.showVolume : true
-        },
-        price_line() {
-            return 'priceLine' in this.sett ?
-                this.sett.priceLine : true
-        },
-        colorCandleUp() {
-            return this.sett.colorCandleUp ||
-            this.$props.colors.candleUp
-        },
-        colorCandleDw() {
-            return this.sett.colorCandleDw ||
-            this.$props.colors.candleDw
-        },
-        colorWickUp() {
-            return this.sett.colorWickUp ||
-            this.$props.colors.wickUp
-        },
-        colorWickDw() {
-            return this.sett.colorWickDw ||
-            this.$props.colors.wickDw
-        },
-        colorWickSm() {
-            return this.sett.colorWickSm ||
-            this.$props.colors.wickSm
-        },
-        colorVolUp() {
-            return this.sett.colorVolUp ||
-            this.$props.colors.volUp
-        },
-        colorVolDw() {
-            return this.sett.colorVolDw ||
-            this.$props.colors.volDw
+const draw = (ctx) => {
+
+            // If data === main candlestick data
+            // render as main chart:
+            if (props.sub === props.data) {
+                var cnv = {
+                    candles: props.layout.candles,
+                    volume: props.layout.volume,
+                }
+            // Else, as offchart / onchart indicator:
+            } else {
+                cnv = layout_cnv(props)
+            }
+
+            if (show_volume) {
+                var cv = cnv.volume
+                for (var i = 0, n = cv.length; i < n; i++) {
+                    new Volbar(props, ctx, cv[i])
+                }
+            }
+
+            var cc = cnv.candles
+            for (var i = 0, n = cc.length; i < n; i++) {
+                new Candle(props, ctx, cc[i])
+            }
+
+            if (price_line) priceState.value.draw(ctx)
         }
-    },
-    data() {
-        return { price: {} }
-    }
 
+// computed
+const sett = computed(()=> props.settings) // isn't it already reactive ?
+// nested computed - expensive computing
+const show_volume = computed(()=> 'showVolume' in sett ? sett.showVolume : true )
+const price_line = computed(()=> 'priceLine' in this.sett ? this.sett.priceLine : true)
+const colorCandleUp = computed(()=> sett.colorCandleUp || props.colors.candleUp )
+const colorCandleDw = computed(()=> sett.colorCandleDw || props.colors.candleDw )
+const colorWickUp = computed(()=> sett.colorWickUp || props.colors.wickUp )
+const colorWickDw = computed(()=> sett.colorWickDw || props.colors.wickDw )
+const colorWickSm = computed(()=> sett.colorWickSm || props.colors.wickSm )
+const colorVolUp = computed(()=> sett.colorVolUp || props.colors.volUp)
+const colorVolDw = computed(()=> sett.colorVolDw || props.colors.volDw)
+
+export default {
+    name: 'Candles',
+    mixins: [Overlay]
 }
+
 </script>