+++ webpack/dev.config.js
@@ -11,6 +11,11 @@ module.exports = (env, options) => ({
     entry: './src/main.js',
     module: {
         rules: [{
+                test: /\.ts$/, // Regex to test for .ts files
+                use: 'ts-loader', // Use ts-loader for these files
+                exclude: /node_modules/, // Exclude node_modules directory
+            },
+            {
                 test: /\.vue$/,
                 exclude: /node_modules/,
                 loader: 'vue-loader'
@@ -30,7 +35,8 @@ module.exports = (env, options) => ({
             {
                 test: /script_ww\.js$/,
                 loader: 'worker-loader'
-            }
+            },
+            
         ]
     },
     plugins: [